+++ src/main/kotlin/graphql/dataTransferObjects/CannotSetBroadcast.kt
@@ -1,0 +1,7 @@
+package com.neelkamath.omniChatBackend.graphql.dataTransferObjects
+
+import com.neelkamath.omniChatBackend.graphql.routing.Placeholder
+
+object CannotSetBroadcast {
+    fun getPlaceholder() = Placeholder
+}